backend/requirements.txt
# WEB FRAMEWORK
fastapi
uvicorn[standard]
python-multipart
pydantic

# DATABASE
sqlalchemy
psycopg2-binary
alembic

# AUTH & SECURITY
python-jose[cryptography]
passlib[bcrypt]

# ENVIRONMENT
python-dotenv

# OCR & IMAGE PROCESSING
pytesseract
easyocr
opencv-python
pillow
numpy

# PDF PROCESSING
pdf2image
PyMuPDF

# AI & LLM
google-generativeai
google-genai

# TASK QUEUE 
celery
redis
flower

# UTILITIES
requests
tqdm
httpx==0.26.0
boto3
pinecone-client
scikit-learn
